Refactor iRacing module: Consolidate helper functions and remove redundant files
- Moved all helper functions from `helper.ts` to `index.ts` for better organization.
- Updated imports across the codebase to reflect the new structure.
- Removed `helper.ts` file as it is no longer needed.
- Introduced `api.ts`, `authentication.ts`, `cache.ts`, and `transforms.ts` for modular functionality.
- Added detailed JSDoc comments for better documentation and understanding of functions.
- Implemented error handling and logging improvements in API calls.
- Ensured backward compatibility while enhancing the modular design of the iRacing server utilities.
