+++ src/app/(dashboard)/schedule/page.tsx
@@ -1,16 +1,15 @@
-import { ScheduleView } from "@/modules/schedule/ui/views/schedule-view";
-import * as helper from "@/modules/iracing/server/helper";
-import { auth } from "@/lib/auth";
 import { redirect } from "next/navigation";
 import { headers } from "next/headers";
+
 import { DateTime } from "luxon";
+
+import { ScheduleView } from "@/modules/schedule/ui/views/schedule-view";
+import { auth } from "@/lib/auth";
 
 const SchedulePage = async () => {
   const session = await auth.api.getSession({ headers: await headers() });
 
   if (!session) redirect("/sign-in");
-
-  // await helper.cacheSeriesImage2();
 
   const dt = DateTime.now().setZone("America/New_York");
   const resetDay = dt
@@ -24,7 +23,7 @@ const SchedulePage = async () => {
     .plus({
       day: 7,
     });
-    
+
   console.log("Date:", dt.day);
   console.log("Reset Day:", resetDay.day);
