+++ src/modules/iracing/server/index.ts
@@ -1,0 +1,30 @@
+/**
+ * @fileoverview Barrel export for iRacing server utilities
+ *
+ * This module provides a centralized export point for all iRacing server-side
+ * utilities, maintaining backward compatibility while enabling modular organization.
+ *
+ * @example
+ * ```typescript
+ * // Import specific functions
+ * import { getOrRefreshAuthCode, fetchData } from '@/modules/iracing/server';
+ *
+ * // Or import everything (maintains backward compatibility)
+ * import * as helper from '@/modules/iracing/server';
+ * ```
+ */
+
+// Configuration constants
+export * from "./config";
+
+// Authentication utilities
+export * from "./authentication";
+
+// API utilities
+export * from "./api";
+
+// Caching utilities
+export * from "./cache";
+
+// Data transformation utilities
+export * from "./transforms";
